Include preceding char in reserved word matching
Check the preceding char when looking for matches to reserved words,
NULL, TRUE, FALSE, NA, Inf, and NaN. Without this check, these words
will be highlighed if they are the last letters of a longer word. For
example, the characters in 'TRUE' and 'FALSE' are colored when 'isTRUE'
or 'isFALSE' is in the output.

Add the isword() function to check that the preceding and following
characters are ASCII letters.